[Web] Fix : :fire: move room switch modal to SideBar organism #328
사이드바 컴포넌트 가장 상단에 모달이 나타나야 하기 때문에 기존 위치에서 SideBar 컴포넌트로 모달 컴포넌트 이동
